+++ app/src/main/java/woowacourse/shopping/data/di/RepositoryModule.kt
@@ -1,10 +1,12 @@
 package woowacourse.shopping.data.di
 
 import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.CartRepositoryImpl
 import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.ProductRepositoryImpl
 
 class RepositoryModule {
-    fun provideProductRepository(): ProductRepository = ProductRepository()
+    fun provideProductRepository(): ProductRepository = ProductRepositoryImpl()
 
-    fun provideCartRepository(): CartRepository = CartRepository()
+    fun provideCartRepository(): CartRepository = CartRepositoryImpl()
 }
